code cleanup, readme update, mqtt login popup is not on the app login screen anymore
