+++ README.md
@@ -1,0 +1,26 @@
+# Dashboard
+the data Dashboard for Hi-Horizon racing team. 
+
+For more information, check the [wiki](https://github.com/Hi-Horizon/Dashboard/wiki)
+
+## Running the application
+NOTE: For the application to run, it is required to have Node.js and npm installed.
+
+Clone/download the repository onto your machine. before running the applications install all the dependencies by navigating to ```/webapp``` in your terminal and run the command:
+```
+ npm install
+```
+Once all the dependencies have been downloaded you can run the file ```Hi-horizon_telemetry.bat``` to launch the app.
+
+
+<!-- ## installation
+clone the repository into your system and run the following commands to install the dependencies
+```
+pip install paho-mqtt
+npm install
+```
+then you can host the website locally with:
+```
+npm run dev
+```
+And then run the python script MqttDataHandler.py to connect to the Mqtt broker -->
