--- webapp/README.md
@@ -1,16 +1,0 @@
-# Dashboard
-the data Dashboard for Hi-Horizon racing team
-
-## installation
-clone the repository into your system and run the following commands to install the dependencies
-```
-pip install paho-mqtt
-npm install
-```
-then you can host the website locally with:
-```
-npm run dev
-```
-And then run the python script MqttDataHandler.py to connect to the Mqtt broker
-
-for more information, check the [wiki](https://github.com/Hi-Horizon/Dashboard/wiki)
