--- webapp/src-tauri/2
@@ -1,12 +1,0 @@
-
-added 2 packages, and audited 277 packages in 9s
-
-39 packages are looking for funding
-  run `npm fund` for details
-
-4 low severity vulnerabilities
-
-To address all issues (including breaking changes), run:
-  npm audit fix --force
-
-Run `npm audit` for details.
